change int datatype into float datatype

+++ cell02/ex00/iszero.py
@@ -1,7 +1,7 @@
 def is_zero(n: str) -> bool:
     """Check if the input string represents the integer zero."""
     try:
-        return int(n) == 0
+        return float(n) == 0
     except (ValueError, TypeError):
         return False
